Ensure putSnapshot path honoring case insensitive contract (#85)
## Summary

This is a bug fixes for put-snapshot code path, where UUID-extraction
process is using `tableId` and `databaseId` from the request itself.
Those ids, if directly obtained from top-level request body, can lost
casing if the requests were from platform like Spark SQL. Since the
underlying storage (e.g. HDFS ) is case sensitive in its path URL, we
will need to ensure the original casing info when issue the first commit
as part of CTAS is preserved in the process of UUID-extraction of the
second commit.

The other parts in this PR is to ensure, when `put` is happening, the
`tableDto` is not always built from scratch when there's existing object
discovered by `findById` method previously. This is done by switch
`orElse` method to `orElseGet`, in which the latter will only call the
supplier lazily when the calling object is absent. This leads to a
wasteful implementation as well as confusion on the code stack.

This PR also include:
- Some refactoring to share code with existing code. 
- Ensure `text-fixtures` module has its repository horning case
insensitive contract so that this behavior can be tested through
embedded instances.
- Testing the casing contract in both SQL and Catalog API.