Fix github build failure (#137)
## Summary

The build failed because of missing `khttp` library that is required by
`springdoc-openapi` plugin. This library has been removed from maven2
repo from 2022, so not sure why the previous build succeeded (maybe
github caches some gradle libraries). The `springdoc-openapi` plugin has
removed the usage of `khttp` too since `1.5.0`. I'm upgrading the
version to `1.6.0` which is the latest that gradle 6 supports.

More details in this issue:
https://github.com/springdoc/springdoc-openapi-gradle-plugin/pull/92.

## Changes

- [ ] Client-facing API Changes
- [ ] Internal API Changes
- [x] Bug Fixes
- [ ] New Features
- [ ] Performance Improvements
- [ ] Code Style
- [ ] Refactoring
- [ ] Documentation
- [ ] Tests

For all the boxes checked, please include additional details of the
changes made in this pull request.

## Testing Done
<!--- Check any relevant boxes with "x" -->

- [ ] Manually Tested on local docker setup. Please include commands
ran, and their output.
- [ ] Added new tests for the changes made.
- [ ] Updated existing tests to reflect the changes made.
- [x] No tests added or updated. Please explain why. If unsure, please
feel free to ask for help.
- [ ] Some other form of testing like staging or soak time in
production. Please explain.

For all the boxes checked, include a detailed description of the testing
done for the changes made in this pull request.

# Additional Information

- [ ] Breaking Changes
- [ ] Deprecations
- [ ] Large PR broken into smaller PRs, and PR plan linked in the
description.

For all the boxes checked, include additional details of the changes
made in this pull request.