Refactoring auth init and adding tests (#133)
## Summary
problem: 1) auth requirement is not documented in swagger UI 2) auth
interceptor init is not tested

(for context, this PR is from a ticket that is meant for me to onboard
to the team and learn the code. as part of skilling up I've used
refactoring to learn the code better. also, a lot of this code was made
in an attempt to add custom headers as types exposed in our client SDK
which turned out to be more complexity than it is worth.)

changes:
- refactored client codegen shell scripts to use a parameterized
openapi-codegen cli, unblocking version upgrade (bumped version to 6.6.0
- logic for converting from clusterproperties yaml -> java objects is
pulled out from TablesMvc and into its own util class for testability,
refactored for readability.
- - along with move and refactoring,  3 unittests to prevent regression
- SwaggerConfig.java deleted and merged into TablesMvcConfigurer to both
reduce # of places to look for config as well as utilize shared
initialization steps. swagger will now init with security documented in
UI if security is declared in cluster properties.
- Authorization is now required in e2e/h2/* tests which enables testing
that the springboot app initializes properly with auth interceptor. new
unittests for this
- Automocked MVC will now include Authorization header by default,
enabling unittests to use auth required endpoints and without any extra
configuration
- moved AuthorizationPropertiesInitializer to be used in either h2 or
mock locations. small unittest for this too
- AuthorizationPropertiesInitializer now also defines auth interceptor

still TODO:

- [x] fix naming of getCons (handlerinterceptor class loader)
- [x] unittest the new HandlerLoader logic
- [x] add documentation in this PR
- [x] investigate whether front end docs + back end validation can be
consolidated further

## Changes

- [ ] Client-facing API Changes
- [ ] Internal API Changes
- [ ] Bug Fixes
- [ ] New Features
- [ ] Performance Improvements
- [ ] Code Style
- [ ] Refactoring
- [ ] Documentation
- [ ] Tests

For all the boxes checked, please include additional details of the
changes made in this pull request.

## Testing Done
<!--- Check any relevant boxes with "x" -->

- [ ] Manually Tested on local docker setup. Please include commands
ran, and their output.
- [x] Added new tests for the changes made.
- [ ] Updated existing tests to reflect the changes made.
- [ ] No tests added or updated. Please explain why. If unsure, please
feel free to ask for help.
- [ ] Some other form of testing like staging or soak time in
production. Please explain.

added several unittests:
- covering clusterpropertiesutil 
- for proving interceptor initialization is correct and includes headers
when needed
- some small tests for loading clusterproperties from yaml for use in
unittests
- as well as deployed locally to verify that swagger had auth documented
on endpoints


# Additional Information

- [ ] Breaking Changes
- [ ] Deprecations
- [ ] Large PR broken into smaller PRs, and PR plan linked in the
description.

For all the boxes checked, include additional details of the changes
made in this pull request.