Adding time based filter to exclude recently created tables from jobs (#163)
## Summary

<!--- HINT: Replace #nnn with corresponding Issue number, if you are
fixing an existing issue -->

[Issue](https://github.com/linkedin/openhouse/issues/#nnn)] Briefly
discuss the summary of the changes made in this
pull request in 2-3 lines.
JobsScheduler considers tables which are recently created also for
running maintenance jobs. This has two concerns:
1. Newly created tables seldom require maintenance hence launching jobs
for them is un-neccesary
2. Some accounts have table creation process followed by immediate
deletion lead to race conditions where jobs scheduler `gets` a table but
during spark job execution, the table may be deleted leading to job
failures. Testing flows from downstream systems is one example which can
lead to this case.

Solution:
Only consider tables which have spent significant time in Openhouse
catalog.

Implementation:
Added integer `timeFilter` option in JobsScheduler cmd options which
translates to hours before which a table should be created to be
considered for Job runs.

## Changes

- [ ] Client-facing API Changes
- [ ] Internal API Changes
- [x] Bug Fixes
- [x] New Features
- [ ] Performance Improvements
- [ ] Code Style
- [ ] Refactoring
- [ ] Documentation
- [ ] Tests

For all the boxes checked, please include additional details of the
changes made in this pull request.

## Testing Done
<!--- Check any relevant boxes with "x" -->

- [ ] Manually Tested on local docker setup. Please include commands
ran, and their output.
- [x] Added new tests for the changes made.
- [ ] Updated existing tests to reflect the changes made.
- [ ] No tests added or updated. Please explain why. If unsure, please
feel free to ask for help.
- [ ] Some other form of testing like staging or soak time in
production. Please explain.

For all the boxes checked, include a detailed description of the testing
done for the changes made in this pull request.

# Additional Information

- [ ] Breaking Changes
- [ ] Deprecations
- [ ] Large PR broken into smaller PRs, and PR plan linked in the
description.

For all the boxes checked, include additional details of the changes
made in this pull request.