Added a regex based db and table name storage selector (#191)
## Summary

Introduced an implementation of Storage Selector that accepts a regex
and storage-type and selects the provided storage type if db and table
match the regex pattern. if pattern does not match default storage is
used.

## Changes

- [ ] Client-facing API Changes
- [ ] Internal API Changes
- [ ] Bug Fixes
- [x] New Features
- [ ] Performance Improvements
- [ ] Code Style
- [ ] Refactoring
- [ ] Documentation
- [x] Tests

For all the boxes checked, please include additional details of the
changes made in this pull request.

## Testing Done
<!--- Check any relevant boxes with "x" -->

- [x] Manually Tested on local docker setup. Please include commands
ran, and their output.
- [x] Added new tests for the changes made.
- [x] Updated existing tests to reflect the changes made.
- [ ] No tests added or updated. Please explain why. If unsure, please
feel free to ask for help.
- [ ] Some other form of testing like staging or soak time in
production. Please explain.

For all the boxes checked, include a detailed description of the testing
done for the changes made in this pull request.
* Ran oh-s3-spark docker-compose with 2 configured storages (s3 & hdfs)
with default as s3 and hdfs for tables matching the regex
".*prod.*\.table.*"

![Screenshot 2024-09-14 at 9 41
02 AM](https://github.com/user-attachments/assets/9df86a76-a4ed-44d6-9528-5e145e8308b6)

* Created 2 tables 
* "database.table1" that doesnt match regex and defaults to s3 storage
* "prod_database.table1" that matches regex and chooses hdfs as storage
![Screenshot 2024-09-14 at 9 42
01 AM](https://github.com/user-attachments/assets/f1d5042c-4d10-46db-a409-8428d7db8b17)

* Verified selected storages in hts using mysql
![Screenshot 2024-09-14 at 9 42
30 AM](https://github.com/user-attachments/assets/c0539ab9-087b-456a-b507-ed0fba4a7cd2)

# Additional Information

- [ ] Breaking Changes
- [ ] Deprecations
- [ ] Large PR broken into smaller PRs, and PR plan linked in the
description.

For all the boxes checked, include additional details of the changes
made in this pull request.