Support wildcard based matching for togglestatus endpoint (#203)
This PR contains logically two components: 
- The changes in
`iceberg/openhouse/internalcatalog/src/main/java/com/linkedin/openhouse/internal/catalog/OpenHouseInternalCatalog.java`
is due to the discovery in the testing phase that the an inheritance
doesn't apply for annotation and AOP's proxy model requires the
annotation to be applied on the object that got directly invoked. Thus
the annotation in this `createTable` is not necessary. Since adding
annotation was the only reason why rewriting this method, removing this
whole method doesn't change any behavior.
- The main part of this PR is to add wildcard matching for
`hts/togglestatus` endpoint, so that for features like temporarily
forbidding table-creation, we can simply insert one rule that contains
wildcard for both tableId and databaseId.